updated rename controllers folder
